Remove DGKernelBase (#10114)
